VimL: only include function tags generated by regexp
'man ctags' says the following about '--regex-<LANG>' option:
"When a matching line is found, a tag will be generated for the name
defined by replacement [...]."
Note that a new tag is generated, but the old one is not replaced by it.

Thus 'neoinclude' will twice include tags for functions matching the
regexp specified by '--regex-vim' option. The first tag is generated by
ctags itself, and the second tag is generated by regexp.

To avoid this problem 'ctags' command should be invoked with the
additional '--vim-kinds=-f' option to omit function tags that were not
generated by regexp.